+++ libs/ett/security/routing-security/src/index.ts
@@ -1,2 +1,2 @@
-export { ettAuthRoutePath } from './lib/ett-security-route-path';
-export { EttAuthenticationGuard } from './lib/ett-security-authentication.guard';
+export { ettAuthRoutePath } from './lib/ett-auth-route-path';
+export { EttAuthenticationGuard } from './lib/ett-authentication.guard';
