+++ libs/ett/security/routing-security/src/lib/ett-authentication.guard.spec.ts
@@ -1,19 +1,18 @@
-import { Location } from '@angular/common';
-import { Component } from '@angular/core';
-import { TestBed } from '@angular/core/testing';
-import { Router } from '@angular/router';
-import { AuthOidcQueryParameterName } from '@energinet-datahub/ett/auth/data-access-api';
-import { ettAuthRoutePath } from '@energinet-datahub/ett/security/routing-security';
-import { SpectacularAppComponent } from '@ngworker/spectacular';
-import { render, RenderResult } from '@testing-library/angular';
-
-import { EttAuthenticationGuard } from './ett-security-authentication.guard';
-
 describe(EttAuthenticationGuard.name, () => {
   @Component({
     template: '',
   })
   class TestGuardedComponent {}
+  import { Location } from '@angular/common';
+  import { Component } from '@angular/core';
+  import { TestBed } from '@angular/core/testing';
+  import { Router } from '@angular/router';
+  import { AuthOidcQueryParameterName } from '@energinet-datahub/ett/auth/data-access-api';
+  import { SpectacularAppComponent } from '@ngworker/spectacular';
+  import { render, RenderResult } from '@testing-library/angular';
+
+  import { ettAuthRoutePath } from './ett-auth-route-path';
+  import { EttAuthenticationGuard } from './ett-authentication.guard';
 
   @Component({
     template: '',
